Version 2.3.7 - replace deprecated gdk function / gkrellmd error flush
Use gdk_pixbuf_new_from_data() instead of deprecated
gdk_pixbuf_new_from_inline().  Inline alert_inline[] needed to be --raw.

When gkrellmd refused connections, the socket was not flushed and the error
message might not be sent back to the client.  Also, add a dialog to display
the error message.