Avoid corruption of chart labels
Do not try to render text that is not a valid utf8 string. This may
happen when formatting the chart text resulted in garbage which
apparently may happen for very huge I/O speeds, at least in older
GKrellM releases. Whether or not this was fixed by changing IO counters
to 64bit values is not entirely clear (and I lack a system to reproduce
this right now).

Change based on gkrellm-2.3.5-fix-diskio-corruption.patch originally
written by Stefan Seyfried, thanks for the patch.