server: add debug logging to allow host logic
Add debug logging to see which of the allow-host rules from
gkrellmd.conf matched or did not match a connecting client. This makes
it possible to debug typos in the configuration or bugs in the allow
list handling.

I used this debug output to debug issue #13 (CIDR match not working for
IPv6 addresses).